pyproject.toml
[tool.poetry]
name = "line-bot"
version = "0.1.0"
description = ""
authors = ["Mishima-Ryuji <ryuji.mishima.com@gmail.com>"]
readme = "README.md"
packages = [{include = "line_bot"}]

[tool.poetry.dependencies]
python = "^3.9"
pylint = "^2.15.4"
python-dotenv = "^0.21.0"
Flask = "^2.2.2"
notebook = "^6.5.1"
firebase-admin = "^6.0.0"
gunicorn = "^20.1.0"
line-bot-sdk = "^2.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
